test: fold space key remount scenario
